% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchablePDF.R
\name{convert_pdf_searchable}
\alias{convert_pdf_searchable}
\title{Batch Convert PDFs into searchable PDF}
\usage{
convert_pdf_searchable(
  google_auth_file_loc,
  input_directory,
  pdf_export_loc,
  python_interpreter = NULL
)
}
\arguments{
\item{google_auth_file_loc}{This is the file path of your Google Cloud API service account json file}

\item{input_directory}{Folder location that contains the list of PDFs you want to convert}

\item{pdf_export_loc}{Folder location for converted PDF}

\item{python_interpreter}{This is the python interpreter location to be used to create the virtual environment.
By default this is set to NULL. If NULL is set, Python 3.8.7 will be downloaded to your machine in order to create the virtual environment.}
}
\value{

}
\description{
This function converts a list of PDFs into searchable PDF using GoogleVisionAPI
}
\details{
To obtain the authentication json file.
Navigate to it via: Google Dev Console > Credentials > New credentials > Service account Key > Select service account > Key type = JSON

Download the JSON file.
}
\examples{
\donttest{

google_auth_loc <- 'your google api service authentication.json'

pdf_export_loc = '/Users/jenny/pdf_output'

input_directory <- 'your input directory'

# example if you dont have python installed and want to use the automated downloaded one (3.8.7)
convert_pdf_searchable(google_auth_loc,python_interpreter,input_directory,pdf_export_loc)

# example if you already have python and you want to use that instead
convert_pdf_searchable(google_auth_loc,python_interpreter = 'usr/bin/python',input_directory,pdf_export_loc)

}
}
