% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python.R
\name{setup_python_virtual_env}
\alias{setup_python_virtual_env}
\title{Setup python virtual environment for SearchablePDF}
\usage{
setup_python_virtual_env(python_interpreter = NULL, force_installation = FALSE)
}
\arguments{
\item{python_interpreter}{This is the python interpreter location to be used to create the virtual environment.
By default this is set to NULL. If NULL is set, Python 3.8.7 will be downloaded to your machine in order to create the virtual enviornment.}

\item{force_installation}{This will force to recreate the virtual environment}
}
\value{
virtual environment path information
}
\description{
This function creates a virtual environment that SearchablePDF will run on. You can provide the your existing Python interpreter location if you want to use that.
}
\details{
The main function convert_pdf_searchable will automatically run a validation to check if you have the correct virtual env setup and install it by running this function.
}
\examples{
\dontrun{
# example to download a new Python if you dont already have one.
setup_python_virtual_env()

# example to use the existing Python version
setup_python_virtual_env(python_interpreter = 'usr/bin/python')



}





}
