% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchablePDF.R
\name{convert_pdf_searchable}
\alias{convert_pdf_searchable}
\title{Batch Convert PDFs into searchable PDF}
\usage{
convert_pdf_searchable(google_auth_file_loc, input_directory, pdf_export_loc)
}
\arguments{
\item{google_auth_file_loc}{This is the file path of your Google Cloud API service account json file}

\item{input_directory}{Folder location that contains the list of PDFs you want to convert}

\item{pdf_export_loc}{Folder location for converted PDF}
}
\description{
This function converts a list of PDFs into searchable PDF using GoogleVisionAPI
}
\details{
To obtain the authentication json file.
Navigate to it via: Google Dev Console > Credentials > New credentials > Service account Key > Select service account > Key type = JSON

Download the JSON file.
}
\examples{
\dontrun{

google_auth_loc <- 'your google api service authentication.json'

pdf_export_loc = '/Users/jenny/pdf_output'

input_directory <- 'your input directory'

# example if you dont have python installed and want to use the automated
downloaded one (3.8.7)
setup_python_virtual_env()
convert_pdf_searchable(google_auth_loc,input_directory,pdf_export_loc)


# example if you already have python and you want to use that instead
setup_python_virtual_env(python_interpreter = 'usr/bin/python')
convert_pdf_searchable(google_auth_loc,input_directory,pdf_export_loc)


# example if you want to change the default virutalenv path into something
else
setup_python_virtual_env(python_interpreter = 'usr/bin/python',
 virtual_path = 'my/path/to/.virtualenv')
convert_pdf_searchable(google_auth_loc,input_directory,pdf_export_loc)

}
}
